pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["roaddust2 <roaddust2@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
pytest-cov = "^4.0.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
validators = "^0.20.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
